Yet Another Hack to make translations work on dashboard
When you try to start a translation without assigning
it first, it sends the language of the task on the
blocked_languages list and that breaks everything.

Since we assign the task on a ajax call, make that
ajax call an utils function and then use it on 
the perform_task_js file so we can assign before
opening the dialog/starting the review/approve.

Yeah, sorry.